+++ web/src/js/enroute.js
@@ -1,17 +1,6 @@
-var ADD_TIME = "Add Time";
-var ADD_DIST = "Add Distance";
-var CROSS_TIME = "Cross Time";
-
-
-/**
- *
-    dT  = delta time - the delay in minutes
-    dNm = separation distance gained
-    t   = target time
-    DTR = distance to run
- *
- */
-
+const ADD_TIME = "Add Time";
+const ADD_DIST = "Add Distance";
+const CROSS_TIME = "Cross Time";
 const STANDARD_MSL_TEMPERATURE = 15;
 const SPECIFIC_HEAT_RATIO = 1.4;
 const SPECIFIC_GAS_CONSTANT = 287;
@@ -22,6 +11,13 @@ class SpeedValues {
         this.groundSpeed = groundSpeed;
         this.machNumber = machNumber;
         this.changedValue = changedValue;
+    }
+}
+
+class VectorValues {
+    constructor(heading, time) {
+        this.heading = heading;
+        this.time = time
     }
 }
 
@@ -38,7 +34,7 @@ class EnrouteCalculator {
 
     validateValues() {
         if (this.groundSpeed > 100 && this.flightLevel < 650 &&
-            this.distanceToFix > 100 && (this.reportedMach < 5 || this.reportedMach === 0) & !isNaN(this.dataPoint)
+            this.distanceToFix >= 50 && (this.reportedMach < 5 || this.reportedMach === 0) & !isNaN(this.dataPoint)
         ) {
             return true;
         } else {
@@ -117,8 +113,7 @@ class EnrouteCalculator {
 
     prevailingWinds() {
         let trueAirSpeed = this.reportedMach * EnrouteCalculator.calculateSpeedOfSound(this.flightLevel);
-        let winds = this.groundSpeed - trueAirSpeed;
-        return winds;
+        return this.groundSpeed - trueAirSpeed;
     }
 
     // Vector to Fix (Time);
@@ -127,45 +122,43 @@ class EnrouteCalculator {
     // Total Distance = Initial Distance + Extra Distance;
     // Vector to Fix (Cross Time, If > Original)
     // Same as VTF (Time);
-
     vectorToFixAddTime() {
-        let results = [];
-        let totalDistance = this.distanceToFix + (this.dataPoint * this.groundSpeed);
-        [10, 20, 30, 40].forEach(function (x) {
-            results.push(EnrouteCalculator.vectorToFix(x, totalDistance));
-        });
-
-        return results;
+        let groundSpeedPerMinute = this.groundSpeed / 60;
+        let extraDistance = this.dataPoint * groundSpeedPerMinute;
+        return EnrouteCalculator.vectorToFix(this.distanceToFix, extraDistance, this.groundSpeed);
 	}
 
 	vectorToFixAddDistance() {
-        let results = [];
-        let totalDistance = this.distanceToFix + this.dataPoint;
-        [10, 20, 30, 40].forEach(function (x) {
-            results.push(EnrouteCalculator.vectorToFix(x, totalDistance));
-        });
-
-        return results;
+        return EnrouteCalculator.vectorToFix(this.distanceToFix, this.dataPoint, this.groundSpeed);
 	}
 
 	vectorToFixCrossTime() {
-        let results = [];
         let initialTimeToCross = (this.distanceToFix / this.groundSpeed) * 60;
         let groundSpeedPerMinute = this.groundSpeed / 60;
         let extraDistance = (this.dataPoint - initialTimeToCross) * groundSpeedPerMinute;
-        [10, 20, 30, 40].forEach(function (x) {
-            if (this.dataPoint < initialTimeToCross) {
-                return [-1, -1, -1, -1];
-            } else {
-                EnrouteCalculator.vectorToFix(x, extraDistance);
-            }
-        });
-
-        return results;
+        return EnrouteCalculator.vectorToFix(this.distanceToFix, extraDistance, this.groundSpeed);
 	}
 
-	static vectorToFix(headingDegree, extraDistance) {
+	// This is an isoceles triangle.
+	// https://www.mathsisfun.com/algebra/trig-solving-sss-triangles.html
+    // https://owlcation.com/stem/Everything-About-Triangles-and-More-Isosceles-Equilateral-Scalene-Pythagoras-Sine-and-Cosine
+	static vectorToFix(initialDistance, extraDistance, groundSpeed) {
+        let totalDistance = extraDistance + initialDistance;
+        let time = (totalDistance / (groundSpeed / 60)) / 2;
+        let a = (totalDistance) / 2;
+        let b = a;
+        let c = initialDistance;
 
+        let cosB = ((c * c) + (a * a) - (b * b)) / (2 * c * a);
+        let radiansB = Math.acos(cosB);
+        let vector = radiansB * (180 / Math.PI);
+        //let roundedVector = vector - (vector % 10);
+
+        //if ((vector % 10) > 5) {
+        //    roundedVector += 10;
+        //}
+
+        return new VectorValues(vector, time);
     }
 
     vectorToReturnAddTime() {
@@ -212,11 +205,11 @@ class EnrouteCalculator {
 class EnrouteUI {
      static calculate() {
         if (ENROUTE_CALCULATOR.validateValues()) {
-            if (ENROUTE_CALCULATOR.currentMode == ADD_TIME) {
+            if (ENROUTE_CALCULATOR.currentMode === ADD_TIME) {
                 return EnrouteUI.addTime();
-            } else if (ENROUTE_CALCULATOR.currentMode == ADD_DIST) {
+            } else if (ENROUTE_CALCULATOR.currentMode === ADD_DIST) {
                 return EnrouteUI.addDistance();
-            } else if (ENROUTE_CALCULATOR.currentMode == CROSS_TIME) {
+            } else if (ENROUTE_CALCULATOR.currentMode === CROSS_TIME) {
                 return EnrouteUI.crossTime();
             } else {
                 return -1;
@@ -227,22 +220,43 @@ class EnrouteUI {
     static addTime() {
         let speedValues = ENROUTE_CALCULATOR.applySpeedTime();
         EnrouteUI.writeApplySpeed(speedValues);
+
+        if (speedValues.changedValue > 0) {
+            let vectorValues = ENROUTE_CALCULATOR.vectorToFixAddTime();
+            EnrouteUI.writeVectorDirect(vectorValues);
+        } else {
+            EnrouteUI.resetVectorDirectOutputs();
+        }
     }
 
     static addDistance() {
         let speedValues = ENROUTE_CALCULATOR.applySpeedDistance();
         EnrouteUI.writeApplySpeed(speedValues);
+        let vectorValues = ENROUTE_CALCULATOR.vectorToFixAddDistance();
+        EnrouteUI.writeVectorDirect(vectorValues);
     }
 
     static crossTime() {
         let speedValues = ENROUTE_CALCULATOR.applySpeedCrossTime();
         EnrouteUI.writeApplySpeed(speedValues);
+
+        if (speedValues.changedValue > 0) {
+            let vectorValues = ENROUTE_CALCULATOR.vectorToFixCrossTime();
+            EnrouteUI.writeVectorDirect(vectorValues);
+        } else {
+            EnrouteUI.resetVectorDirectOutputs();
+        }
     }
 
     static writeApplySpeed(speedValues) {
          $("#enr_out_mach_number").html(speedValues.machNumber.toFixed(2));
          $("#enr_out_ground_speed").html(Math.round(speedValues.groundSpeed));
          $("#enr_out_data_point").html(Math.round(speedValues.changedValue));
+    }
+
+    static writeVectorDirect(vectorValues) {
+         $("#enr_out_vector_fix_degrees").html(Math.round(vectorValues.heading) + "&#176;");
+         $("#enr_label_out_vector_fix_time").html(Math.round(vectorValues.time) + " mins");
     }
 
     static fieldChange() {
@@ -269,15 +283,15 @@ class EnrouteUI {
         $("#enr_in_data_point").val("");
         $(".enr_out").html("&nbsp;-&nbsp;");
 
-        if (theButton.val() == ADD_TIME) {
+        if (theButton.val() === ADD_TIME) {
             $("#enr_label_data_point").html("Delay Required");
             $("#enr_current_mode").val(ADD_TIME);
             EnrouteUI.switchOutDataPointUnit("+Separation: ", "NM");
-        } else if (theButton.val() == ADD_DIST) {
+        } else if (theButton.val() === ADD_DIST) {
             $("#enr_label_data_point").html("Gap Required");
             $("#enr_current_mode").val(ADD_DIST);
             EnrouteUI.switchOutDataPointUnit("+Time: ", "mins");
-        } else if  (theButton.val() == CROSS_TIME) {
+        } else if  (theButton.val() === CROSS_TIME) {
             $("#enr_label_data_point").html("Time to Cross");
             $("#enr_current_mode").val(CROSS_TIME);
             EnrouteUI.switchOutDataPointUnit("+Separation: ", "NM");
@@ -302,6 +316,10 @@ class EnrouteUI {
     static resetOutputs() {
          $(".enr_out").html("&nbsp;-&nbsp;");
          $("#enr_out_data_point").html("");
+    }
+
+    static resetVectorDirectOutputs() {
+         $(".enr_out_vector_fix").html("&nbsp;-&nbsp;");
     }
 }